Migrate DittoAndroidTools to Ditto Kotlin SDK v5

+++ DittoToolsAndroid/src/main/java/live/ditto/tools/diskusage/DiskUsageState.kt
@@ -2,7 +2,7 @@ package live.ditto.tools.diskusage
 
 data class DiskUsageState(
     val rootPath: String = "ditto",
-    val totalSizeInBytes: Int = 0,
+    val totalSizeInBytes: Long = 0L,
     val totalSize: String = "Calculating...",
     val children: List<DiskUsage> = listOf(DiskUsage()),
     val unhealthySizeInBytes: Int = FIVE_HUNDRED_MEGABYTES_IN_BYTES
