+++ DittoToolsAndroid/src/main/java/live/ditto/tools/diskusage/usecase/GetDiskUsageMetrics.kt
@@ -15,10 +15,10 @@ class GetDiskUsageMetrics {
 
     fun execute(currentState: DiskUsageState): HealthMetric {
 
-        val dittoStoreSize: Int =
-            currentState.children.firstOrNull { shortRelativePath(it.relativePath) == DITTO_STORE }?.sizeInBytes ?: 0
-        val dittoReplicationSize: Int = currentState.children.firstOrNull {
-            shortRelativePath(it.relativePath) == DITTO_REPLICATION }?.sizeInBytes ?: 0
+        val dittoStoreSize: Long =
+            currentState.children.firstOrNull { shortRelativePath(it.relativePath) == DITTO_STORE }?.sizeInBytes ?: 0L
+        val dittoReplicationSize: Long = currentState.children.firstOrNull {
+            shortRelativePath(it.relativePath) == DITTO_REPLICATION }?.sizeInBytes ?: 0L
 
         val isHealthy = healthCheckSize(dittoStoreSize, dittoReplicationSize)
 
@@ -36,7 +36,7 @@ class GetDiskUsageMetrics {
         )
     }
 
-    private fun healthCheckSize(dittoStoreSize: Int, dittoReplicationSize: Int): Boolean {
+    private fun healthCheckSize(dittoStoreSize: Long, dittoReplicationSize: Long): Boolean {
         val totalBytes = dittoStoreSize + dittoReplicationSize
 
         return totalBytes <= unhealthySizeInBytes
